% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_filepath.R
\name{fix_filepath}
\alias{fix_filepath}
\title{Fix a copied windows filepath}
\usage{
fix_filepath()
}
\value{
Character vector that has a cleaned, modified windows file path replacing
all forward slashes with a single backslash so it can be recopied and pasted in
windows file explorer
}
\description{
Fix a copied windows filepath
}
\examples{
\dontrun{
clipr::write_clip("C:\\\\Users\\\\Test", allow_non_interactive = TRUE)
fix_filepath()
}
}
