% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_path.R
\name{open_path}
\alias{open_path}
\title{open_path.}
\usage{
open_path()
}
\value{
Open path will read a url, file path, directory label, etc.. from
your clipboard and open that for you. Just note, there's not much safety net
regarding checking for a valid url, file path, etc... because of the
variability that may be associated with them. The function will throw an error
in two cases: not used in an interactive session (ie trying to call from
command line) or if the path cannot be opened.
}
\description{
open_path.
}
\examples{
\dontrun{
clipr::write_clip("C:\\\\Users\\\\Test")
open_path()

clipr::write_clip("https://github.com/cpilat97/fixFilepath")
open_path()
}
}
